refactor: refact SkuSelector component
