+++ packages/core/src/components/ui/SkuSelector/Selectors.tsx
@@ -6,11 +6,12 @@ import {
   SkuSelectorProps,
   SkuOption,
 } from '@faststore/ui'
-import type { SkuVariantsByName } from './SkuSelectorsContext'
-import { SkuSelectorsContext } from './SkuSelectorsContext'
-
 import NextLink from 'next/link'
-import { getSkuSlug } from './skuVariants'
+
+export type SkuVariantsByName = Record<
+  string,
+  Array<{ alt: string; label: string; value: string; src?: string }>
+>
 
 interface Props extends HTMLAttributes<HTMLDivElement> {
   /**
@@ -48,52 +49,23 @@ function Selectors({
   availableVariations,
   ...otherProps
 }: Props) {
-  // `dominantVariation` will be ever the first variant
-  const [dominantVariation] = useMemo(() => {
-    return Object.keys(activeVariations)
-  }, [activeVariations])
-
-  const getItemHref = useCallback(
-    (option: SkuOption) => {
-      const currentItemHref = `/${getSkuSlug(
-        slugsMap,
-        {
-          ...activeVariations,
-          [dominantVariation]: option.value,
-        },
-        dominantVariation
-      )}/p`
-      return currentItemHref
-    },
-    [activeVariations, dominantVariation, slugsMap]
-  )
-
   return (
     <section {...otherProps}>
-      <SkuSelectorsContext.Provider
-        value={{
-          slugsMap,
-          availableVariations,
-          activeVariations,
-          dominantProperty: dominantVariation,
-        }}
-      >
-        {availableVariations &&
-          Object.keys(availableVariations).map((skuVariant) => (
-            <UISkuSelector
-              key={skuVariant}
-              skuPropertyName={skuVariant}
-              options={availableVariations[skuVariant]}
-              ImageComponent={ImageComponent}
-              activeVariations={activeVariations}
-              linkProps={{
-                as: NextLink,
-                legacyBehavior: false,
-              }}
-              getItemHref={getItemHref}
-            />
-          ))}
-      </SkuSelectorsContext.Provider>
+      {availableVariations &&
+        Object.keys(availableVariations).map((skuVariant) => (
+          <UISkuSelector
+            key={skuVariant}
+            skuPropertyName={skuVariant}
+            options={availableVariations[skuVariant]}
+            ImageComponent={ImageComponent}
+            activeVariations={activeVariations}
+            slugsMap={slugsMap}
+            linkProps={{
+              as: NextLink,
+              legacyBehavior: false,
+            }}
+          />
+        ))}
     </section>
   )
 }
